fix: file name ， Folder Internal Component Refactoring
