+++ src/UIComponents/headers/BtnHeader.jsx
@@ -1,7 +1,8 @@
 import React from 'react'
-import style from './HeaderBtn.module.scss'
+import ButtonUI from '../buttons/ButtonUI'
+import style from './BtnHeader.module.scss'
 
-export default function HeaderText() {
+export default function HeaderBtn({ title, text }) {
   return (
     <div className={style.header__container}>
       <div className={style.header__return}>
@@ -12,9 +13,12 @@ export default function HeaderText() {
         </div>
         <div className={style.text}>
           <div className={style.text__container}>
-            <div className={style.title}>推文</div>
+            <div className={style.title}>{title}</div>
           </div>
         </div>
+      </div>
+      <div className={style.header__btn}>
+        <ButtonUI btnStyle="btn__pill__middle" text={text} />
       </div>
     </div>
   )
