+++ src/UIComponents/headers/TextHeader.module.scss
@@ -3,26 +3,23 @@
 
 .header__container {
   width: 100%;
-  height: 100%;
   display: flex;
   align-items: center;
+  border-bottom: 1px solid color.$assistant-gray-color;
 
   .header__return {
     width: 100%;
-    height: 100%;
     display: flex;
     gap: 1rem;
     align-items: center;
-
+    padding: 1rem 1.5rem;
     .icon {
-      width: fit-content;
-      padding: 0.3125rem 0.1875rem 0.3125rem 0.25rem;
-
+      cursor: pointer;
       .icon__container {
         display: flex;
         align-items: center;
-        width: 1.5rem;
-        height: 1.5rem;
+        width: 1.0625rem;
+        height: 0.875rem;
 
         .img {
           content: url('../../public/arrow@2x.png');
@@ -33,8 +30,6 @@
     }
 
     .text {
-      height: 100%;
-
       .text__container {
         .title {
           @include mixin.font-setting(1.5rem, 700, color.$gray-10);
