+++ src/UIComponents/headers/UserHeader.module.scss
@@ -3,21 +3,18 @@
 
 .header__container {
   width: 100%;
-  height: 100%;
   display: flex;
   gap: 1rem;
   align-items: center;
-
+  padding: 1rem 1.5rem;
+  border-bottom: 1px solid color.$assistant-gray-color;
   .icon {
-    width: fit-content;
-    height: 100%;
-    padding: 0.3125rem 0.1875rem 0.3125rem 0.25rem;
-
+    cursor: pointer;
     .icon__container {
       display: flex;
       align-items: center;
-      width: 1.5rem;
-      height: 1.5rem;
+      width: 1.0625rem;
+      height: 0.875rem;
 
       .img {
         content: url('../../public/arrow@2x.png');
@@ -28,7 +25,6 @@
   }
 
   .text {
-    height: 100%;
     .text__container {
       .name {
         @include mixin.font-setting(1.125rem, 700, color.$gray-10);
